backend/visuals/requirements.txt
django
djangorestframework
asgiref
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
psycopg2-binary
python-dotenv
markdown 
django-filter
django_import_export